Adição carrossel de imagens Recursividade no repositório para buscar todas as infos dos pokemons e suas variedades Dialogo para descrição das habilidades

+++ app/src/main/java/com/chenriquevz/pokedex/ui/home/HomeFragment.kt
@@ -103,7 +103,7 @@ class HomeFragment : Fragment(), Injectable {
             override fun onQueryTextSubmit(query: String): Boolean {
 
                 if (query.isNotEmpty()) {
-                    val action = HomeFragmentDirections.homeToPokemon(query.toInt())
+                    val action = HomeFragmentDirections.homeToPokemon(query)
                     navController.navigate(action)
                 } else {
                     _context.toast(getString(R.string.search_error))
